cleanup back office contract
