+++ accounting-domain/src/main/java/onlydust/com/marketplace/accounting/domain/port/in/AccountingRewardPort.java
@@ -1,6 +1,5 @@
 package onlydust.com.marketplace.accounting.domain.port.in;
 
-import onlydust.com.marketplace.accounting.domain.model.Invoice;
 import onlydust.com.marketplace.accounting.domain.model.RewardId;
 import onlydust.com.marketplace.accounting.domain.view.BackofficeRewardView;
 import onlydust.com.marketplace.kernel.model.RewardStatus;
@@ -11,14 +10,10 @@ import java.util.List;
 
 public interface AccountingRewardPort {
 
-    List<BackofficeRewardView> findByInvoiceId(Invoice.Id invoiceId);
-
     Page<BackofficeRewardView> getRewards(int pageIndex, int pageSize,
                                           List<RewardStatus> statuses,
                                           Date fromRequestedAt, Date toRequestedAt,
                                           Date fromProcessedAt, Date toProcessedAt);
-
-    List<BackofficeRewardView> processingRewardsByInvoiceIds(List<Invoice.Id> invoiceIds);
 
     String exportRewardsCSV(List<RewardStatus> statuses,
                             Date fromRequestedAt, Date toRequestedAt,
