+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/it/bo/BackOfficeAccountingApiIT.java
@@ -669,26 +669,6 @@ public class BackOfficeAccountingApiIT extends AbstractMarketplaceBackOfficeApiI
 
         // When
         client.post()
-                .uri(getApiURI(POST_REWARDS_SEARCH))
-                .header("Api-Key", apiKey())
-                .contentType(APPLICATION_JSON)
-                .bodyValue("""
-                            {
-                                "invoiceIds": ["%s"]
-                            }
-                        """.formatted(invoiceId))
-                // Then
-                .exchange()
-                .expectStatus()
-                .is2xxSuccessful()
-                .expectBody()
-                .jsonPath("$.rewards.length()").isEqualTo(1)
-                .jsonPath("$.rewards[0].id").isEqualTo(rewardId.toString())
-                .jsonPath("$.rewards[0].status").isEqualTo("PROCESSING")
-                .jsonPath("$.rewards[0].paymentId").doesNotExist();
-
-        // When
-        client.post()
                 .uri(getApiURI(POST_REWARDS_PAY.formatted(rewardId)))
                 .header("Api-Key", apiKey())
                 .contentType(APPLICATION_JSON)
@@ -702,24 +682,6 @@ public class BackOfficeAccountingApiIT extends AbstractMarketplaceBackOfficeApiI
                 .exchange()
                 .expectStatus()
                 .isNoContent();
-
-        // When
-        client.post()
-                .uri(getApiURI(POST_REWARDS_SEARCH))
-                .header("Api-Key", apiKey())
-                .contentType(APPLICATION_JSON)
-                .bodyValue("""
-                            {
-                                "invoiceIds": ["%s"]
-                            }
-                        """.formatted(invoiceId))
-                // Then
-                .exchange()
-                .expectStatus()
-                .is2xxSuccessful()
-                .expectBody()
-                .jsonPath("$.rewards.length()").isEqualTo(0);
-
 
         client.get()
                 .uri(getApiURI(String.format(ME_REWARD, rewardId)))
