+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/repository/RewardDetailsViewRepository.java
@@ -109,20 +109,6 @@ public interface RewardDetailsViewRepository extends JpaRepository<BackofficeRew
                                                                Pageable pageable);
 
     @Query(value = SELECT + """
-            where (coalesce(:invoiceStatuses) is null or cast(i.status as text) in (:invoiceStatuses))
-            and (coalesce(:invoiceIds) is null or i.id in (:invoiceIds))
-            and (coalesce(:rewardStatuses) is null or cast(rs.status as text) in (:rewardStatuses))
-            """, nativeQuery = true)
-    List<BackofficeRewardViewEntity> findAllByInvoiceStatusesAndInvoiceIdsAndRewardStatuses(List<String> invoiceStatuses, List<UUID> invoiceIds,
-                                                                                            List<String> rewardStatuses);
-
-    @Query(value = SELECT + """
-            where i.id = :invoiceId
-            """, nativeQuery = true)
-    List<BackofficeRewardViewEntity> findAllByInvoiceId(@NonNull UUID invoiceId);
-
-
-    @Query(value = SELECT + """
             where r.id in (:rewardIds)
             """, nativeQuery = true)
     List<BackofficeRewardViewEntity> findAllByRewardIds(@NonNull List<UUID> rewardIds);
